fix mod to page structure
