+++ platform/src/routes.rs
@@ -4,8 +4,8 @@ use crate::pages::attributes::AttributePage;
 use crate::pages::create::CreatePage;
 use crate::pages::dashboard::DashboardPage;
 use crate::pages::find_email::FindEmailPage;
-use crate::pages::groups::GroupPage;
-use crate::pages::groups::_id::GroupDetailPage;
+use crate::pages::groups::_id::page::GroupDetailPage;
+use crate::pages::groups::page::GroupPage;
 use crate::pages::id::response_report::ResponseReportPage;
 use crate::pages::id::select_response::response_type::SelectResponseDetailPage;
 use crate::pages::id::select_response::SelectResponsePage;
@@ -13,8 +13,8 @@ use crate::pages::id::survey_summary::SurveySummaryPage;
 use crate::pages::id::write_question::WriteQuestionPage;
 use crate::pages::id::write_title::WriteTitlePage;
 use crate::pages::login::LoginPage;
-use crate::pages::members::MemberPage;
-use crate::pages::members::_id::MemberDetailPage;
+use crate::pages::members::_id::page::MemberDetailPage;
+use crate::pages::members::page::MemberPage;
 use crate::pages::reset_password::ResetPasswordPage;
 use crate::prelude::*;
 use dioxus_translate::Language;
